inference/requirements.txt
azure-identity
azure-keyvault-secrets
azure-storage-blob
pandas
Flask
scikit-learn
